Remove bilinear filtering
Change the scaling method to the Qt default of nearest neighbor. Change
scale factor floats to doubles. Restructure rendering pipeline so that
the QPainter performs the scaling rather than scaling the pixmap.

+++ main.cpp
@@ -57,10 +57,10 @@ int main(int argc, char *argv[])
             &ntr, SLOT(initStream()));
     QObject::connect(&w, SIGNAL(sendNfcPatch(int)),
             &ntr, SLOT(writeNFCPatch(int)));
-    QObject::connect(&w, SIGNAL(topScaleChanged(float)),
-            &top, SLOT(setScale(float)));
-    QObject::connect(&w, SIGNAL(botScaleChanged(float)),
-            &bot, SLOT(setScale(float)));
+    QObject::connect(&w, SIGNAL(topScaleChanged(double)),
+            &top, SLOT(setScale(double)));
+    QObject::connect(&w, SIGNAL(botScaleChanged(double)),
+            &bot, SLOT(setScale(double)));
 
     QObject::connect(&ntr, SIGNAL(streamStarted()),
             t, SLOT(start()));
